Add SeparationUnit class which simulates an hypothetical separation
This unit simulates a separation based on component splits. This unit does not have an impact on CAPEX or OPEX since it is an hypothetical separation to study its impact on the process. Its main use is to study the impact of solvent recovery on the MSP though uncertainty and sensitivity analyses, for example.